strip_end関数のユニットテストを追加

+++ test/unit/strip_end/cmd.c
@@ -4,7 +4,7 @@
 int main(){
     char s[][10] = {"abc\n", "abc    ", "abc\t", "abc\t\n", "abc"};
 
-    for(int i = 0; i < sizeof(s) / sizeof(s[0]); i++) {
+    for(size_t i = 0; i < sizeof(s) / sizeof(s[0]); i++) {
         printf("%s:\n", strip_end(s[i]));
     }
     return 0;
